+++ core-ng/src/main/java/core/framework/log/message/PerformanceStat.java
@@ -7,9 +7,9 @@ import core.framework.api.json.Property;
  */
 public class PerformanceStat {
     @Property(name = "total_elapsed")
-    public Long totalElapsed;
+    public Long totalElapsed = 0L;
     @Property(name = "count")
-    public Integer count;
+    public Integer count = 0;
     @Property(name = "read_entries")
     public Integer readEntries;
     @Property(name = "write_entries")